Make metric value to long to avoid too big integer issue. (#1876)
* Make metric value to long to avoid too big integer issue.

* Update protocol.

* Add a new filter to service number.

+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2AggregationQueryDAO.java
@@ -102,7 +102,7 @@ public class H2AggregationQueryDAO implements IAggregationQueryDAO {
                 while (resultSet.next()) {
                     TopNEntity topNEntity = new TopNEntity();
                     topNEntity.setId(resultSet.getString(Indicator.ENTITY_ID));
-                    topNEntity.setValue(resultSet.getInt("value"));
+                    topNEntity.setValue(resultSet.getLong("value"));
                     topNEntities.add(topNEntity);
                 }
             } catch (SQLException e) {
